Move server implementation to header
