+++ cpp/CMakeLists.txt
@@ -58,7 +58,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DJSON_DIAGNOSTICS=1")
 
 include_directories("${PROJECT_SOURCE_DIR}/include")
 
-add_executable(example_server src/example_server.cpp src/server.cpp)
+add_executable(example_server src/example_server.cpp)
 target_link_libraries(example_server CONAN_PKG::nlohmann_json CONAN_PKG::websocketpp)
 
 file(GLOB TEST_SOURCES test/*.cpp)
